web scraper/requirements.txt
python-dotenv
serpapi
beautifulsoup4
requests
langchain
langchain-ollama
langchain-community
langchain-core